requirements.txt
# Common packages
briefcase
tomli
tomli-w
django
django-jinja
django-session-security
django-guardian
django-crispy-forms
djangorestframework
crispy-bootstrap5
pendulum
pytest
pytest-django

# Processing
numpy
pandas
openpyxl
scipy
pillow
pydicom
dicom2nifti
python-gdcm
pynetdicom
tensorflow==2.15.0
torch
torchvision
torchsummary
opencv-python
totalsegmentator

# Required for Docker
psycopg2-binary
gunicorn

# To fix warnings requests
requests==2.28.2
urllib3==1.26.14
charset_normalizer==2.1.1